Criação do objeto Perfil, e implementação do CRUD do Perfil. -Gabriel França

+++ PontoX.Domain.Core/Interfaces/Repositories/IRepositoryBase.cs
@@ -6,9 +6,9 @@ namespace PontoX.Domain.Core.Interfaces.Repositories
     public interface IRepositoryBase<T> where T : class
     {
         Task<IEnumerable<T>> Consultar();
-        Task<T> Consultar(int id);
+        Task<T> Consultar(long id);
         Task<bool> Adicionar(T model);
         Task<bool> Atualizar(T model);
-        Task<bool> Remover(int id);
+        Task<bool> Remover(long id);
     }
 }
