+++ PontoX.Domain.Service/ServiceBase.cs
@@ -14,11 +14,11 @@ namespace ServiceBlackDomain.Services
 
         public async Task<bool> Adicionar(T model) => await _repository.Adicionar(model);
 
-        public async Task<bool> Remover(int id) => await _repository.Remover(id);
+        public async Task<bool> Remover(long id) => await _repository.Remover(id);
 
         public async Task<IEnumerable<T>> Consultar() => await _repository.Consultar();
 
-        public async Task<T> Consultar(int id) => await _repository.Consultar(id);
+        public async Task<T> Consultar(long id) => await _repository.Consultar(id);
 
         public async Task<bool> Atualizar(T model) => await _repository.Atualizar(model);
     }
